--- app/controllers/albums_controller.rb
@@ -1,5 +1,0 @@
-class AlbumsController < ApplicationController
-  def index
-    @data = Album.where(artist_id: params['artist_id'])
-  end
-end
