+++ app/controllers/api/v1/artists_controller.rb
@@ -1,0 +1,9 @@
+module Api
+  module V1
+    class ArtistsController < ApplicationController
+      def index
+        @data = Artist.all
+      end
+    end
+  end
+end
