--- app/controllers/artists_controller.rb
@@ -1,5 +1,0 @@
-class ArtistsController < ApplicationController
-  def index
-    @data = Artist.all
-  end
-end
