+++ app/views/api/v1/albums/index.jbuilder
