+++ app/views/api/v1/artists/index.jbuilder
